pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"
[project]
name = "bluesky_nexus"
version = "1.0.0"
description = "Generation of the NeXus file from a blusky run using pydantic schema-model approach."
keywords = ["bluesky", "nexus", "hdf5", "data-science", "pydantic"]
authors = [
    { name = "Daniel Tomecki", email = "daniel.tomecki@helmholtz-berlin.de" },
]
license = { file = "LICENSE" }

requires-python = ">=3.10"
dependencies = [
    "bluesky==1.13",
    "h5py==3.12.1",
    "pydantic==2.10.2",
    "pytz==2024.2",
    "PyYAML==6.0.2",
    "numpy",
    "fonttools>=4.22.0",
    "kiwisolver>=1.3.1",
    "pillow>=8",
    "pyparsing>=2.3.1",
    "types-pytz>=2022.1.1",
    "mkdocs==1.6.1",
    "mkdocs-material==9.5.38",
    "mkdocs-drawio-file==1.5.2",
    "click>=7.0",
    "jinja2>=2.11.1,<4.0.0",
    "markupsafe>=2.0.1",
    "beautifulsoup4>=4.11.0,<5.0.0",
    "requests>=2.27.1,<3.0.0",
    "babel~=2.10",
    "colorama~=0.4",
]
[project.optional-dependencies]
dev = ["pytest>=8.3"]
